refactor(config): rework config helpers
