+++ libs/dynamic-api/src/helpers/index.ts
@@ -1,2 +1,5 @@
-export * from './config.helper';
 export * from './route-decorators.helper';
+export * from './route-description.helper';
+export * from './swagger-config.helper';
+export * from './validation-config.helper';
+export * from './versioning-config.helper';
