+++ libs/dynamic-api/src/helpers/validation-config.helper.ts
@@ -1,0 +1,14 @@
+import {
+  INestApplication,
+  ValidationPipe,
+  ValidationPipeOptions,
+} from '@nestjs/common';
+
+
+function enableDynamicAPIValidation(app: INestApplication, options: ValidationPipeOptions = {}) {
+  app.useGlobalPipes(
+    new ValidationPipe(options),
+  );
+}
+
+export { enableDynamicAPIValidation };
