+++ libs/dynamic-api/src/helpers/versioning-config.helper.ts
@@ -1,0 +1,13 @@
+import { INestApplication, VersioningOptions, VersioningType } from '@nestjs/common';
+
+function enableDynamicAPIVersioning(
+  app: INestApplication,
+  options?: VersioningOptions,
+) {
+  app.enableVersioning({
+    type: VersioningType.URI,
+    ...options,
+  });
+}
+
+export { enableDynamicAPIVersioning };
