+++ libs/dynamic-api/src/interfaces/index.ts
@@ -1,6 +1,6 @@
 export * from './entity-mappers.interface';
 export * from './dynamic-api-options.interface';
 export * from './dynamic-api-schema-options.interface';
-export * from './maybe-promise.type';
+export * from './dynamic-api-swagger-options.type';
 export * from './route-config.interface';
 export * from './service-provider.interface';
